requirements.txt
Django==4.0.7
psycopg2-binary==2.9.3
django-environ==0.9.0
django-js-routes==0.2.0
inertia-django==0.3.2
django-webpack-loader==1.6.0
pytest==7.1.2
pytest-django==4.5.2
pytest-cov==3.0.0
flake8==5.0.4
black==22.6.0
isort==5.10.1
marshmallow==3.17.1
django-model-utils==4.2.0
pyotp==2.6.0
cryptography==37.0.4
PyQRCode==1.2.1
pypng==0.0.21
mnemonic==0.20
gunicorn==20.1.0
gevent==22.10.2
django-countries==7.3.2
django-auditlog==2.1.0
six==1.16.0
jsonfield==3.1.0
python-dateutil==2.8.2
django-user-agents==0.4.0
requests==2.28.1
celery==5.2.7
redis==4.3.4
flower==1.2.0
django-celery-beat==2.3.0
pytest-xdist[psutil]==2.5.0
django-extensions==3.2.0
PyYAML==6.0